Update Database.cpp
Done following changes in Node class. 
Constructor Initialization List: The constructor now uses an initialization list to directly initialize label and name, which is more efficient.

Const-Correctness: The getProperty and printProperties functions are marked as const, 
indicating that they do not modify the state of the object.

added some methods:

Optional Property Retrieval: The getProperty method allows users to retrieve a specific property by key. 
It returns an empty string if the property does not exist.

deleteProperty:
This method takes a property key as an argument and checks if the property exists in the properties map.
If it does, the property is removed, and a message confirming the deletion is printed.
If the property does not exist, an appropriate message is displayed.

clearProperties:
This method clears all properties from the properties map using the clear() function.
After clearing, a message indicating that all properties have been removed is printed.


Improved Comments: I've added and refined comments to make the purpose of each member variable and function clearer. 
to aids readability and maintainability.

added formatting logic to ensure the properties are printed neatly in JSON format.


also removed content of interpretQuery to change formate.